Created All Messages page with a blank HTML table to be populated from the Database.
Renamed current allmessages page to downloadmessages, changed URL link, Description and Navigation bar to reflect the new naming convention.

Added in Monolog logging in downloadmessages.php file, logging is not being saved in the Database.
