+++ includes/coursework/app/templates/banner.html.twig
@@ -4,6 +4,7 @@
     <div id="banner-div">
         <ul>
             <li><a href="{{ landing_page }}">Home</a></li>
+            <li><a href="/coursework_public/downloadmessages">Download Messages</a></li>
             <li><a href="/coursework_public/allmessages">All Messages</a></li>
         </ul>
         <h1>{{ page_heading_1 }}</h1>
