+++ includes/coursework/app/templates/downloadmessages.html.twig
@@ -3,7 +3,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <div id="page-content-div">
         <table class="border" width="100%" align="center">
-            <caption>Table showing all messages stored in the database with their content & metadata</caption>
+            <caption>Table showing downloaded messages with their content & metadata</caption>
             <thead>
             <tr>
                 <th>Source Number</th>
