+++ includes/coursework/app/templates/homepageform.html.twig
@@ -3,6 +3,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <div id="page-content-div">
         <p> Welcome to our Telemetry Data Processing site.</p>
-        <p> Please click on the All Messages button above on the navigation bar</p>
+        <p> Please click on the Download Messages button above on the navigation bar</p>
     </div>
 {% endblock %}
